Wrap useNotificationCount query and favicon changing into sub component so it doesn't cause full app re-renders on every refetch.
See
https://www.developerway.com/posts/why-custom-react-hooks-could-destroy-your-app-performance
for a description of the issue and why this resolves it
