+++ src/features/colinks/CoLinksNav.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import { useContext, useEffect, useState } from 'react';
+import { memo, useContext, useEffect, useState } from 'react';
 
 import { CoLogoMark } from 'features/nav/CoLogoMark';
 import { client } from 'lib/gql/client';
@@ -367,7 +367,7 @@ const NavItem = ({
   );
 };
 
-const Count = () => {
+const Count = memo(function Count() {
   const { count } = useNotificationCount();
 
   return count ? (
@@ -388,4 +388,4 @@ const Count = () => {
       {count}
     </Text>
   ) : null;
-};
+});
